+++ ai_bundle/app.py
@@ -1,3 +1,4 @@
+"""Initializing the app."""
 from fastapi import FastAPI
 
 app = FastAPI(