+++ tests/test_app.py
@@ -1,0 +1,7 @@
+def test_root():
+    """Testing root function."""
+    from fastapi import FastAPI
+
+    from ai_bundle.app import app
+
+    assert isinstance(app, FastAPI)