+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
 <!--scriv-insert-here-->
+
+<a id='changelog-0.1.1'></a>
+# 0.1.1 — 2022-09-18
+
+## Added
+
+- Bringing the coverage to 100% on a basic testing function.
+
+<a id='changelog-0.1.1'></a>
+# 0.1.1 — 2022-09-17
+
+## Added
+
+- Tests and Tox environment
 # 0.1.0 — 2022-09-17
 
 ## Added